Refactor Flutter app structure
